Go standard directory layout
