+++ assets/cluster_small.json
@@ -1,0 +1,11 @@
+{
+    "Id": 1,
+    "Nodes": [
+        {
+            "Id": 1,
+            "Memory": 10,
+            "Cores": [1,2,3,4,5,6,7,8,9,10],
+            "MemoryAvailable": 10,
+            "CoresAvailable": 10
+        }]
+}
