--- cluster_big.json
@@ -1,11 +1,0 @@
-{
-    "Id": 1,
-    "Nodes": [
-        {
-            "Id": 1,
-            "Memory": 100,
-            "Cores": [1,2,3,4,5,6,7,8,9,10],
-            "MemoryAvailable": 100,
-            "CoresAvailable": 100
-        }]
-}
