--- registry/server.go
@@ -1,217 +1,0 @@
-package registry
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"sync"
-	"time"
-)
-
-// Registry server constant port and URL
-const ServerPort = ":3000"
-const ServicesURL = "http://localhost" + ServerPort + "/services"
-
-type registry struct {
-	registrations []Registration
-	mutex         *sync.RWMutex
-}
-
-func (r *registry) add(reg Registration) error {
-	r.mutex.Lock()
-	r.registrations = append(r.registrations, reg)
-	r.mutex.Unlock()
-	err := r.sendRequiredServices(reg)
-	r.notify(patch{
-		Added: []patchEntry{
-			patchEntry{Name: reg.ServiceName,
-				URL: reg.ServiceURL},
-		},
-	})
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-	return nil
-}
-
-func (r *registry) notify(fullPatch patch) {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-
-	for _, reg := range r.registrations {
-		go func(reg Registration) {
-			for _, reqService := range reg.RequiredServices {
-				p := patch{Added: []patchEntry{}, Removed: []patchEntry{}}
-				sendUpdate := false
-				// check if the any added services are required for this service
-				for _, added := range fullPatch.Added {
-					if added.Name == reqService {
-						p.Added = append(p.Added, added)
-						sendUpdate = true
-					}
-				}
-
-				// check if the any removed services are required for this service
-				for _, removed := range fullPatch.Removed {
-					if removed.Name == reqService {
-						p.Removed = append(p.Removed, removed)
-						sendUpdate = true
-					}
-				}
-				if sendUpdate {
-					err := r.sendPatch(p, reg.ServiceUpdateURL)
-					if err != nil {
-						log.Println(err)
-						return
-					}
-				}
-
-			}
-		}(reg)
-	}
-}
-
-func (r *registry) sendRequiredServices(reg Registration) error {
-	r.mutex.RLock()
-	defer r.mutex.RUnlock()
-
-	var p patch
-	for _, serviceReg := range r.registrations {
-		for _, reqService := range reg.RequiredServices {
-			if serviceReg.ServiceName == reqService {
-				p.Added = append(p.Added, patchEntry{
-					Name: serviceReg.ServiceName,
-					URL:  serviceReg.ServiceURL,
-				})
-			}
-		}
-	}
-	err := r.sendPatch(p, reg.ServiceUpdateURL)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (r *registry) sendPatch(p patch, url string) error {
-	d, err := json.Marshal(p)
-	if err != nil {
-		return err
-	}
-	_, err = http.Post(url, "application/json", bytes.NewBuffer(d))
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (r *registry) remove(url string) error {
-	for i := range r.registrations {
-		if r.registrations[i].ServiceURL == url {
-			r.notify(patch{
-				Removed: []patchEntry{
-					patchEntry{
-						Name: r.registrations[i].ServiceName,
-						URL:  r.registrations[i].ServiceURL,
-					},
-				},
-			})
-			r.mutex.Lock()
-			r.registrations = append(r.registrations[:i], r.registrations[i+1:]...)
-			r.mutex.Unlock()
-			return nil
-		}
-	}
-	return fmt.Errorf("Service at URL %v not found", url)
-}
-
-func (r *registry) heartbeat(freq time.Duration) {
-	for {
-		var wg sync.WaitGroup
-		for _, reg := range r.registrations {
-			wg.Add(1)
-			go func(reg Registration) {
-				defer wg.Done()
-				success := true
-				for attempts := 0; attempts < 3; attempts++ {
-					res, err := http.Get(reg.HeartbeatURL)
-					if err != nil {
-						log.Println(err)
-					} else if res.StatusCode == http.StatusOK {
-						log.Printf("Heartbeat check passed for %v", reg.ServiceName)
-						// failed in a previous attempt
-						if !success {
-							r.add(reg)
-						}
-						break
-					}
-					log.Printf("Heartbeat check failed for %v", reg.ServiceName)
-					if success {
-						success = false
-						r.remove(reg.ServiceURL)
-					}
-					time.Sleep(1 * time.Second)
-				}
-			}(reg)
-			wg.Wait()
-			time.Sleep(freq)
-		}
-	}
-}
-
-var once sync.Once
-
-func SetupRegistryService() {
-	once.Do(func() {
-		go reg.heartbeat(3 * time.Second)
-	})
-
-}
-
-var reg = registry{registrations: make([]Registration, 0), mutex: new(sync.RWMutex)}
-
-type RegistryService struct{}
-
-// ServeHTTP implements http.Handler.
-func (r *RegistryService) ServeHTTP(w http.ResponseWriter, req *http.Request) {
-	log.Println("Request received")
-	switch req.Method {
-	case http.MethodPost:
-		dec := json.NewDecoder(req.Body)
-		var r Registration
-		err := dec.Decode(&r)
-		if err != nil {
-			log.Println(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-		log.Printf("Adding service: %v with URL %v\n", r.ServiceName, r.ServiceURL)
-		err = reg.add(r)
-		if err != nil {
-			log.Println(err)
-			w.WriteHeader(http.StatusBadRequest)
-			return
-		}
-	case http.MethodDelete:
-		payload, err := io.ReadAll(req.Body)
-		if err != nil {
-			log.Println(err)
-			return
-		}
-		url := string(payload)
-		log.Printf("Removing service at URL: %v", url)
-		err = reg.remove(url)
-		if err != nil {
-			log.Println(err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-	default:
-		w.WriteHeader(http.StatusMethodNotAllowed)
-		return
-	}
-}
